% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare.R
\docType{package}
\name{declare-package}
\alias{declare}
\alias{declare-package}
\title{declare: Optional type annotation}
\description{
Allows user to add type annotation to a function via comments.
    The package uses a simple mechanism, and it is designed such that the 
    annotation may be "switched on and off" as one pleases. This is an 
    experiment of programming in R with stronger type.
}
\keyword{internal}
