% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_type_helper.R
\name{match_args}
\alias{match_args}
\title{Match a list of arguments against the formal arguments of a function}
\usage{
match_args(ls_args, chr_name)
}
\arguments{
\item{ls_args}{A list of arguments (named or nameless)}

\item{chr_name}{Character vector; the formal arguments of a function}
}
\description{
Match a list of arguments against the formal arguments of a function
}
\keyword{internal}
