% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_with_type_annotation}
\alias{source_with_type_annotation}
\title{Source file with type annotation}
\usage{
source_with_type_annotation(fpath, local = FALSE, show_order = F)
}
\arguments{
\item{fpath}{Character; the full file path.}

\item{local}{TRUE, FALSE or an environment; determining where the parsed
expressions are evaluated. FALSE (the default) corresponds to the user's
workspace (the global environment) and TRUE to the environment from which
source is called.}

\item{show_order}{T or F; if T, print out the expression in the order they
are evaluated.}
}
\description{
Source file with type annotation
}
