% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_type.R
\name{type}
\alias{type}
\title{Declare type of a function}
\usage{
type(f, ...)
}
\arguments{
\item{f}{A function; the function to which type-annotations are added.}

\item{...}{The type / classes for each argument.}
}
\description{
A wrapper for `declare_ls`; just to provide another interface.
}
\examples{
\dontrun{
# Example 1
add_one <- function(x) { x + 1 }
add_one <- declare::type(add_one, x = "numeric")
add_one(10)          # Runs fine
add_one(nrow(1:10))  # Expect type error since nrow of a vector returns NULL.
}
}
